requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
django==5.2.4
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
sqlparse==0.5.3
whitenoise==6.9.0
psycopg2-binary==2.9.10
psycopg==3.2.9
django-cors-headers==4.7.0
gunicorn==23.0.0
packaging==25.0
pip==25.1.1
python-decouple==3.8
python-dotenv==1.1.1
config==0.5.1
dj-database-url==3.0.1
typing-extensions==4.14.1





